WIP: major changes and updates
Collectors:
- improve/implement dynamic form loading
- improve base model loading and parameters
- clean/improve JS code
- rework Frontend views.py to improve collectors' handling
- remove all deprecated collectors' fields in frontend_edit.html template

+++ vulture_os/services/frontend/views.py
@@ -24,9 +24,13 @@ __doc__ = 'Listeners View'
 
 
 # Django system imports
-from django.apps import apps
 from django.conf import settings
-from django.http import (JsonResponse, HttpResponseBadRequest, HttpResponseForbidden, HttpResponseRedirect)
+from django.http import (JsonResponse,
+                         HttpResponse,
+                         HttpResponseServerError,
+                         HttpResponseBadRequest,
+                         HttpResponseForbidden,
+                         HttpResponseRedirect)
 from django.shortcuts import render
 from django.urls import reverse
 from django.utils import timezone
@@ -36,6 +40,7 @@ from django.utils.translation import gettext_lazy as _
 from applications.logfwd.models import LogOM
 from darwin.policy.models import DarwinBuffering, DarwinPolicy
 from gui.forms.form_utils import DivErrorList
+from services.apps import ServicesConfig
 from services.frontend.form import FrontendForm, ListenerForm, LogOMTableForm, FrontendReputationContextForm
 from services.frontend.models import Frontend, FrontendReputationContext, Listener, FILEBEAT_MODULE_CONFIG
 from services.rsyslogd.form import CustomActionsForm, RsyslogConditionForm
@@ -246,9 +251,12 @@ def frontend_edit(request, object_id=None, api=False):
     node_listeners = dict()
     custom_actions_form = None
     darwin_buffering_needs_refresh = False
+    old_api_collector = None
+    api_collector_form = None
     if object_id:
         try:
             frontend = Frontend.objects.get(pk=object_id)
+            old_api_collector = frontend.api_collector
         except ObjectDoesNotExist:
             if api:
                 return JsonResponse({'error': _("Object does not exist.")}, status=404)
@@ -258,10 +266,21 @@ def frontend_edit(request, object_id=None, api=False):
     empty = {} if api else None
     if hasattr(request, "JSON") and api:
         validate = request.JSON.get('validate', True)
-        form = FrontendForm(request.JSON or {}, instance=frontend, error_class=DivErrorList)
+        form = FrontendForm(request.JSON or empty, instance=frontend, error_class=DivErrorList)
+        if frontend:
+            # TODO ensure form is correctly instanciated and used while loading existing Frontend on GUI
+            api_collector_form = ServicesConfig.api_collectors_get_form(
+                frontend.api_parser_type,
+                old_api_collector,
+                data=request.JSON or empty)
     else:
         validate = request.POST.get('validate', True) not in (False, "false", "False")
         form = FrontendForm(request.POST or empty, instance=frontend, error_class=DivErrorList)
+        if frontend:
+            api_collector_form = ServicesConfig.api_collectors_get_form(
+                frontend.api_parser_type,
+                old_api_collector,
+                data=request.POST or empty)
 
     def render_form(front, **kwargs):
         save_error = kwargs.get('save_error')
@@ -292,25 +311,10 @@ def frontend_edit(request, object_id=None, api=False):
                 reputationctx_form_list.append(FrontendReputationContextForm(instance=r_tmp))
 
         custom_actions_form = kwargs.get('custom_actions_form', CustomActionsForm({'custom_actions': form.initial.get("custom_actions", [])}, auto_id=False))
-        logger.info(f"[FRONTEND RENDER FORM] custom_actions_form: {custom_actions_form}")
 
         filebeat_configs = deepcopy(FILEBEAT_MODULE_CONFIG)
         if front and front.filebeat_module and front.filebeat_config:
             filebeat_configs[front.filebeat_module]=front.filebeat_config
-
-        # logger.info(f"[FRONTEND RENDER FORM] front.api_collector.all(): {front.api_collector.all()}")
-        logger.info(f"[FRONTEND RENDER FORM] front.api_parser_type: {front.api_parser_type}")
-        logger.info(f"[FRONTEND RENDER FORM] front.api_collector: {front.api_collector}")
-
-        # TypeError: Abstract models cannot be instantiated.
-        # api_collectors_generic_form = kwargs.get('api_collectors_generic_form', apps.get_app_config("services").api_collectors_generic_form()(instance=front.api_collector.first()))
-        api_collectors_forms = apps.get_app_config("services").api_collectors_forms()
-        logger.info(f"[FRONTEND RENDER FORM] api_collectors_forms.keys(): {api_collectors_forms.keys()}")
-
-        if front:
-            api_collectors_forms[front.api_parser_type] = kwargs.get('api_collector_form', apps.get_app_config("services").api_collectors_get_form(front.api_parser_type, front.api_collector))
-            logger.info(f"[FRONTEND RENDER FORM] api_collectors_forms[front.api_parser_type]: {api_collectors_forms[front.api_parser_type]}")
-
 
         return render(request, 'services/frontend_edit.html',
                       {'form': form, 'listeners': listener_form_list, 'listener_form': ListenerForm(),
@@ -321,9 +325,8 @@ def frontend_edit(request, object_id=None, api=False):
                        'custom_actions': custom_actions_form,
                        'condition_line_form': RsyslogConditionForm(auto_id=False),
                        'filebeat_module_config': filebeat_configs,
-                       'api_collectors_generic_form': apps.get_app_config("services").api_collectors_generic_form(),
-                       'api_collectors_forms': api_collectors_forms,
-                       'object_id': (frontend.id if frontend else "") or "", **kwargs})
+                       'api_collector_form': api_collector_form,
+                       'object_id': (frontend.pk if frontend else "") or "", **kwargs})
 
     if request.method in ("POST", "PUT"):
         """ Handle JSON formatted listeners """
@@ -444,21 +447,28 @@ def frontend_edit(request, object_id=None, api=False):
         # if not custom_actions_form.is_valid():
         #     form.add_error("custom_actions", custom_actions_form.errors.get("custom_actions", []) if api else custom_actions_form.errors.as_data().get("custom_actions", []))
 
-        if hasattr(request, "JSON") and api:
-            api_collector_form = apps.get_app_config("services").api_collectors_get_form(form.data.get("api_parser_type"), frontend.api_collector, data=request.JSON or empty)
-        else:
-            api_collector_form = apps.get_app_config("services").api_collectors_get_form(form.data.get("api_parser_type"), frontend.api_collector, data=request.POST or empty)
+        # TODO ensure logic is correct for first save, modification with new parser_type, and modification with same parser_type
+        if form.data.get("mode", "") == "log" and form.data.get("listening_mode", "") == "api":
+            if "api_parser_type" not in form.changed_data:
+                # No change of mode, listening_mode or collector type, object can just be modified without deletion
+                old_api_collector = None
+            else:
+                # Create a new form for the new collector type
+                api_collector_form = ServicesConfig.api_collectors_get_form(
+                    form.data.get("api_parser_type"),
+                    None,
+                    (request.JSON if api else request.POST) or empty
+                )
 
-        if not api_collector_form.is_valid():
-            form.add_error(None, api_collector_form.errors.as_json() if api else api_collector_form.errors.as_data().values())
-        api_collector_obj = api_collector_form.save(commit=False)
-        # api_collector_form.use_proxy = api_collectors_generic_form.cleaned_data.get('use_proxy')
+            if not api_collector_form:
+                form.add_error(None, f"Unknown collector {form.data.get('api_parser_type')}")
+            elif not api_collector_form.is_valid():
+                form.add_error(None, api_collector_form.errors.as_json() if api else api_collector_form.errors.as_data().values())
 
         old_nodes = frontend.get_nodes() if frontend else []
         old_rsyslog_filename = frontend.get_rsyslog_base_filename() if frontend and frontend.has_rsyslog_conf else ""
         old_filebeat_filename = frontend.get_filebeat_base_filename() if frontend and frontend.has_filebeat_conf else ""
         old_haproxy_filename = frontend.get_base_filename() if frontend and frontend.has_haproxy_conf else ""
-        old_api_collector = frontend.api_collector
 
         # Frontend used by workflow type change check
         if object_id and "mode" in form.changed_data:
@@ -583,11 +593,17 @@ def frontend_edit(request, object_id=None, api=False):
                                                          arg_field=reputationctx.arg_field,
                                                          dst_field=reputationctx.dst_field)
 
-            """ Delete previous api collector """
-            old_api_collector.delete()
-            api_collector_obj.frontend = frontend
-            logger.debug(f"Saving api_collector {str(api_collector_obj)}")
-            api_collector_obj.save()
+            """ Handle api collector changes """
+            # Having a value here means a collector was linked to this Frontend but doesn't need anymore
+            if old_api_collector:
+                old_api_collector.delete()
+            # This is set only if a collector was set up
+            if api_collector_form:
+                # Can be a previously-existing collector, or a new one
+                api_collector = api_collector_form.save(commit=False)
+                api_collector.frontend = frontend
+                logger.debug(f"Saving api_collector {str(api_collector)}")
+                api_collector.save()
 
             new_nodes = frontend.reload_conf()
             for node in new_nodes:
@@ -723,47 +739,45 @@ def frontend_pause(request, object_id, api=False):
     return JsonResponse({'status': True, 'message': res})
 
 
-def frontend_api_collector_form(request):
+def frontend_api_collector_form(request, collector_name):
     try:
-        type_parser = request.POST.get("api_parser_type")
-        parser_from = apps.get_app_config("services").api_collectors_get_form(type_parser)
-        return JsonResponse({"status": True, "data": str(parser_from)})
+        collector_form = ServicesConfig.api_collectors_get_form(collector_name, data=request.GET or None)
+        return HttpResponse(collector_form.render())
 
     except Exception as e:
-        logger.error(e, exc_info=1)
-        return JsonResponse({
-            'status': False,
-            'error': str(e)
-        })
+        logger.exception(f"Error while getting collector '{collector_name}' form: {e}")
+        return HttpResponseServerError()
 
 
 def frontend_test_apiparser(request):
     try:
-        type_parser = request.POST.get('api_parser_type')
+        collector_type = request.POST.get('api_parser_type')
+        collector_form = ServicesConfig.api_collectors_get_form(collector_name=collector_type, data=request.POST)
+        assert collector_form, f"Unknown collector {collector_type}"
 
-        data = {}
-        for k, v in request.POST.items():
-            if v in ('false', 'true'):
-                v = v == "true"
+        if not collector_form.is_valid():
+            logger.error(f"Collector '{collector_type}' form errors: {collector_form.errors.as_json()}")
+            return JsonResponse({
+                'status': False,
+                'error': "Some fields are not valid",
+                'errors': collector_form.errors.as_json(),
+            }, status=400)
 
-            data[k] = v
+        collector_instance = collector_form.save(commit=False)
+        return JsonResponse(collector_instance.test())
 
-        if data.get('api_parser_use_proxy', True) and data.get('api_parser_custom_proxy', None):
-            # parse_proxy_url will validate and return a correct url
-            proxy = parse_proxy_url(data.get('api_parser_custom_proxy', None))
-            if not proxy:
-                return JsonResponse({'status': False, 'error': "Wrong proxy format"})
-            data['api_parser_custom_proxy'] = proxy
-
-        parser = get_api_parser(type_parser)(data)
-        return JsonResponse(parser.test())
-
-    except Exception as e:
-        logger.error(e, exc_info=1)
+    except AssertionError as e:
         return JsonResponse({
             'status': False,
             'error': str(e)
-        })
+        }, status=404)
+
+    except Exception as e:
+        logger.exception(e)
+        return JsonResponse({
+            'status': False,
+            'error': str(e)
+        }, status=500)
 
 
 def frontend_fetch_apiparser_data(request):
